+++ network-mux/CHANGELOG.md
@@ -8,6 +8,9 @@
 * Polling of the egress queue
 * run, miniProtocolJob, monitor now accept MuxTracerBundle record
   instead of `Tracer m Trace` type.
+* Removed handshake trace events from mux trace.
+* `Trace` was split into three traces: `Trace` of mux events, `ChannelTrace`
+   & `BearerTrace`.  As a result `run` & `Bearer` API were modified.
 
 ### Non-breaking changes
 * Define msHeaderLength instead of using '8'